Removed left pane of tags and moved to menu
I put all the tags we could insert in the menu up above, instead of on the
left pane. It seemed like a good idea in the begginning when we wanted to
do a drag insert, but that may be too much of a hassle so we'll stick with
this for now.